server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.0.1
numpy==1.24.4
tensorflow==2.14.0
